+++ trunk/bpel-g/org.activebpel.rt.bpel.server/src/main/java/org/activebpel/rt/bpel/server/engine/storage/sql/AeQueryRunner.java
@@ -170,7 +170,7 @@ public class AeQueryRunner extends QueryRunner
             {
                AeBlobInputStream stream = (AeBlobInputStream) param;
                // This is binary blob data set the stream
-               aStatement.setBinaryStream(i + 1, stream, (int) stream.length());
+               aStatement.setBinaryStream(i + 1, stream, stream.length());
             }
             else
             {
